@@ -1,23 +1,23 @@
 apply plugin: 'com.android.application'
-apply plugin: 'com.neenbedankt.android-apt'
+apply plugin: 'com.jakewharton.butterknife'
 apply plugin: 'org.greenrobot.greendao'
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
 
     defaultConfig {
         applicationId "wapchief.com.collectiondemo"
         minSdkVersion 17
-        targetSdkVersion 25
+        targetSdkVersion 29
         multiDexEnabled true
         versionCode 2
         versionName "1.0.2"
 
-        ndk {
+//        ndk {
             //选择要添加的对应cpu类型的.so库。
-            abiFilters 'armeabi', 'armeabi-v7a', 'arm64-v8a'
+//            abiFilters 'armeabi', 'armeabi-v7a', 'arm64-v8a'
             // 还可以添加 'x86', 'x86_64', 'mips', 'mips64'
-        }
+//        }
 
         manifestPlaceholders = [
                 JPUSH_PKGNAME : applicationId,
