@@ -1,7 +1,7 @@
 package wapchief.com.collectiondemo.activity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.text.TextPaint;
 import android.view.View;
 import android.widget.Button;
