@@ -1,7 +1,7 @@
 package wapchief.com.collectiondemo.activity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.widget.ImageView;
 
 import com.wapchief.likestarlibrary.like.TCHeartLayout;
