@@ -1,6 +1,6 @@
 package wapchief.com.collectiondemo.fragment;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import java.util.HashMap;
 
