@@ -21,15 +21,15 @@ public class DaoMaster extends AbstractDaoMaster {
 
     /** Creates underlying database table using DAOs. */
     public static void createAllTables(Database db, boolean ifNotExists) {
-        MessageDao.createTable(db, ifNotExists);
         CarShopDao.createTable(db, ifNotExists);
+        MessageDao.createTable(db, ifNotExists);
         UserDao.createTable(db, ifNotExists);
     }
 
     /** Drops underlying database table using DAOs. */
     public static void dropAllTables(Database db, boolean ifExists) {
-        MessageDao.dropTable(db, ifExists);
         CarShopDao.dropTable(db, ifExists);
+        MessageDao.dropTable(db, ifExists);
         UserDao.dropTable(db, ifExists);
     }
 
@@ -49,8 +49,8 @@ public class DaoMaster extends AbstractDaoMaster {
 
     public DaoMaster(Database db) {
         super(db, SCHEMA_VERSION);
-        registerDaoClass(MessageDao.class);
         registerDaoClass(CarShopDao.class);
+        registerDaoClass(MessageDao.class);
         registerDaoClass(UserDao.class);
     }
 
