@@ -8,12 +8,12 @@ import org.greenrobot.greendao.database.Database;
 import org.greenrobot.greendao.identityscope.IdentityScopeType;
 import org.greenrobot.greendao.internal.DaoConfig;
 
-import wapchief.com.collectiondemo.greendao.model.Message;
 import wapchief.com.collectiondemo.greendao.model.CarShop;
+import wapchief.com.collectiondemo.greendao.model.Message;
 import wapchief.com.collectiondemo.greendao.model.User;
 
-import wapchief.com.collectiondemo.greendao.MessageDao;
 import wapchief.com.collectiondemo.greendao.CarShopDao;
+import wapchief.com.collectiondemo.greendao.MessageDao;
 import wapchief.com.collectiondemo.greendao.UserDao;
 
 // THIS CODE IS GENERATED BY greenDAO, DO NOT EDIT.
@@ -25,48 +25,48 @@ import wapchief.com.collectiondemo.greendao.UserDao;
  */
 public class DaoSession extends AbstractDaoSession {
 
-    private final DaoConfig messageDaoConfig;
     private final DaoConfig carShopDaoConfig;
+    private final DaoConfig messageDaoConfig;
     private final DaoConfig userDaoConfig;
 
-    private final MessageDao messageDao;
     private final CarShopDao carShopDao;
+    private final MessageDao messageDao;
     private final UserDao userDao;
 
     public DaoSession(Database db, IdentityScopeType type, Map<Class<? extends AbstractDao<?, ?>>, DaoConfig>
             daoConfigMap) {
         super(db);
 
-        messageDaoConfig = daoConfigMap.get(MessageDao.class).clone();
-        messageDaoConfig.initIdentityScope(type);
-
         carShopDaoConfig = daoConfigMap.get(CarShopDao.class).clone();
         carShopDaoConfig.initIdentityScope(type);
+
+        messageDaoConfig = daoConfigMap.get(MessageDao.class).clone();
+        messageDaoConfig.initIdentityScope(type);
 
         userDaoConfig = daoConfigMap.get(UserDao.class).clone();
         userDaoConfig.initIdentityScope(type);
 
-        messageDao = new MessageDao(messageDaoConfig, this);
         carShopDao = new CarShopDao(carShopDaoConfig, this);
+        messageDao = new MessageDao(messageDaoConfig, this);
         userDao = new UserDao(userDaoConfig, this);
 
-        registerDao(Message.class, messageDao);
         registerDao(CarShop.class, carShopDao);
+        registerDao(Message.class, messageDao);
         registerDao(User.class, userDao);
     }
     
     public void clear() {
-        messageDaoConfig.clearIdentityScope();
         carShopDaoConfig.clearIdentityScope();
+        messageDaoConfig.clearIdentityScope();
         userDaoConfig.clearIdentityScope();
-    }
-
-    public MessageDao getMessageDao() {
-        return messageDao;
     }
 
     public CarShopDao getCarShopDao() {
         return carShopDao;
+    }
+
+    public MessageDao getMessageDao() {
+        return messageDao;
     }
 
     public UserDao getUserDao() {
